Deprecate functional_pil and functional_tensor
