+++ test/test_transforms.py
@@ -10,7 +10,7 @@ import pytest
 import torch
 import torchvision.transforms as transforms
 import torchvision.transforms.functional as F
-import torchvision.transforms.functional_tensor as F_t
+import torchvision.transforms._functional_tensor as F_t
 from PIL import Image
 from torch._utils_internal import get_file_path_2
