+++ torchvision/transforms/v2/functional/_color.py
@@ -4,8 +4,8 @@ import PIL.Image
 import torch
 from torch.nn.functional import conv2d
 from torchvision import datapoints
-from torchvision.transforms import functional_pil as _FP
-from torchvision.transforms.functional_tensor import _max_value
+from torchvision.transforms import _functional_pil as _FP
+from torchvision.transforms._functional_tensor import _max_value
 
 from torchvision.utils import _log_api_usage_once
