+++ torchvision/transforms/v2/functional/_geometry.py
@@ -8,7 +8,7 @@ import torch
 from torch.nn.functional import grid_sample, interpolate, pad as torch_pad
 
 from torchvision import datapoints
-from torchvision.transforms import functional_pil as _FP
+from torchvision.transforms import _functional_pil as _FP
 from torchvision.transforms.functional import (
     _check_antialias,
     _compute_resized_output_size as __compute_resized_output_size,
@@ -19,7 +19,7 @@ from torchvision.transforms.functional import (
     pil_to_tensor,
     to_pil_image,
 )
-from torchvision.transforms.functional_tensor import _pad_symmetric
+from torchvision.transforms._functional_tensor import _pad_symmetric
 
 from torchvision.utils import _log_api_usage_once
